feat(TDI-40568) : TU + fix

+++ core/components-common/src/main/java/org/talend/components/common/tableaction/ConvertAvroTypeToSQL.java
@@ -116,7 +116,7 @@ public class ConvertAvroTypeToSQL {
             sqlType = Types.NUMERIC;
             break;
         case FLOAT:
-            sqlType = Types.FLOAT;
+            sqlType = Types.DOUBLE;
             break;
         case DOUBLE:
             sqlType = Types.DOUBLE;
